Add OmeTiffWriter and MergeOmeTiffWriter classes
