+++ wsireg/writers/ome_tiff_writer.py
@@ -20,9 +20,7 @@ from wsireg.utils.tile_image_transform import (
     subres_zarr_to_tiles,
     tile_pad_output_size,
 )
-from wsireg.utils.tform_utils import (
-    prepare_wsireg_transform_data,
-)
+from wsireg.utils.tform_utils import prepare_wsireg_transform_data
 
 
 class OmeTiffWriter:
@@ -56,13 +54,21 @@ class OmeTiffWriter:
                 self.x_spacing,
             ) = get_final_yx_from_tform(self.reg_image, final_transform)
 
+        self.tile_size = tile_size
+        # protect against too large tile size
+        while (
+            self.y_size / self.tile_size <= 1
+            or self.x_size / self.tile_size <= 1
+        ):
+            self.tile_size = self.tile_size // 2
+
         if by_tile:
             self.y_size, self.x_size = tile_pad_output_size(
-                self.y_size, self.x_size, tile_size=tile_size
+                self.y_size, self.x_size, tile_size=self.tile_size
             )
 
         self.pyr_levels, self.pyr_shapes = get_pyramid_info(
-            self.y_size, self.x_size, self.reg_image.n_ch, tile_size
+            self.y_size, self.x_size, self.reg_image.n_ch, self.tile_size
         )
         self.n_pyr_levels = len(self.pyr_levels)
 
@@ -110,7 +116,7 @@ class OmeTiffWriter:
         compression="default",
     ):
 
-        output_file_name = str(Path(output_dir) / image_name)
+        output_file_name = str(Path(output_dir) / f"{image_name}.ome.tiff")
         self.prepare_image_info(
             image_name,
             final_transform=final_transform,
@@ -121,8 +127,8 @@ class OmeTiffWriter:
 
         rgb_im_data = []
 
-        print(f"saving to {output_file_name}.ome.tiff")
-        with TiffWriter(f"{output_file_name}.ome.tiff", bigtiff=True) as tif:
+        print(f"saving to {output_file_name}")
+        with TiffWriter(output_file_name, bigtiff=True) as tif:
             for channel_idx in range(self.reg_image.n_ch):
                 print(f"transforming : {channel_idx}")
                 if self.reg_image.reader != "sitk":
@@ -157,7 +163,7 @@ class OmeTiffWriter:
                         image = sitk.GetArrayFromImage(image)
 
                     options = dict(
-                        tile=(tile_size, tile_size),
+                        tile=(self.tile_size, self.tile_size),
                         compression=self.compression,
                         photometric="rgb"
                         if self.reg_image.is_rgb
@@ -189,7 +195,7 @@ class OmeTiffWriter:
                                 cv2.INTER_LINEAR,
                             )
                             print(
-                                f"pyr {pyr_idx} : channel {channel_idx} shape: {image.shape}"
+                                f"pyramid index {pyr_idx} : channel {channel_idx} shape: {image.shape}"
                             )
 
                             tif.write(image, **options, subfiletype=1)
@@ -199,7 +205,7 @@ class OmeTiffWriter:
                 rgb_im_data = sitk.GetArrayFromImage(rgb_im_data)
 
                 options = dict(
-                    tile=(tile_size, tile_size),
+                    tile=(self.tile_size, self.tile_size),
                     compression=self.compression,
                     photometric="rgb",
                     metadata=None,
@@ -228,6 +234,7 @@ class OmeTiffWriter:
                             cv2.INTER_LINEAR,
                         )
                         tif.write(rgb_im_data, **options, subfiletype=1)
+        return output_file_name
 
     def write_image_by_tile(
         self,
@@ -242,7 +249,8 @@ class OmeTiffWriter:
         tile_padding=None,
     ):
 
-        output_file_name = str(Path(output_dir) / image_name)
+        output_file_name = str(Path(output_dir) / f"{image_name}.ome.tiff")
+
         self.prepare_image_info(
             image_name,
             final_transform=final_transform,
@@ -257,7 +265,7 @@ class OmeTiffWriter:
         tile_coordinate_data = generate_tile_coords(
             (self.x_size, self.y_size),
             (self.PhysicalSizeX, self.PhysicalSizeY),
-            tile_size=tile_size,
+            tile_size=self.tile_size,
         )
 
         tile_coordinate_data = add_tile_location_on_moving(
@@ -267,10 +275,10 @@ class OmeTiffWriter:
         if tile_padding is None:
             tile_padding = determine_moving_tile_padding(itk_transforms)
 
-        print(f"saving to {output_file_name}.ome.tiff")
-        with TiffWriter(f"{output_file_name}.ome.tiff", bigtiff=True) as tif:
+        print(f"saving to {output_file_name}")
+        with TiffWriter(output_file_name, bigtiff=True) as tif:
             options = dict(
-                tile=(tile_size, tile_size),
+                tile=(self.tile_size, self.tile_size),
                 compression=self.compression,
                 photometric="rgb" if self.reg_image.is_rgb else "minisblack",
                 metadata=None,
@@ -282,7 +290,7 @@ class OmeTiffWriter:
                     self.x_size,
                     tile_coordinate_data,
                     composite_transform,
-                    tile_size=tile_size,
+                    tile_size=self.tile_size,
                     tile_padding=tile_padding,
                 )
                 print(
@@ -308,7 +316,7 @@ class OmeTiffWriter:
                             self.reg_image,
                             resampled_zarray,
                             2 ** pyr_idx,
-                            tile_size,
+                            self.tile_size,
                             self.reg_image.is_rgb,
                         )
                         print(f"pyr {pyr_idx} : RGB-shape: {orig_shape}")
@@ -316,7 +324,7 @@ class OmeTiffWriter:
                         tif.write(
                             subres_zarr_to_tiles(
                                 resampled_zarray_subres,
-                                tile_size,
+                                self.tile_size,
                                 self.reg_image.is_rgb,
                             ),
                             shape=orig_shape,
@@ -333,7 +341,7 @@ class OmeTiffWriter:
                         tile_coordinate_data,
                         composite_transform,
                         ch_idx=channel_idx,
-                        tile_size=tile_size,
+                        tile_size=self.tile_size,
                         tile_padding=tile_padding,
                     )
 
@@ -362,14 +370,14 @@ class OmeTiffWriter:
                                 self.reg_image,
                                 resampled_zarray,
                                 2 ** pyr_idx,
-                                tile_size,
+                                self.tile_size,
                                 self.reg_image.is_rgb,
                             )
 
                             tif.write(
                                 subres_zarr_to_tiles(
                                     resampled_zarray_subres,
-                                    tile_size,
+                                    self.tile_size,
                                     self.reg_image.is_rgb,
                                 ),
                                 shape=orig_shape,
@@ -377,199 +385,4 @@ class OmeTiffWriter:
                                 **options,
                                 subfiletype=1,
                             )
-
-
-class MergeOmeTiffWriter(OmeTiffWriter):
-    def __init__(self, reg_image):
-        super().__init__(reg_image)
-
-    def write_image_by_plane(
-        self,
-        image_name,
-        sub_image_names,
-        transform_data,
-        output_dir="",
-        write_pyramid=True,
-        tile_size=512,
-        compression="default",
-    ):
-
-        if isinstance(sub_image_names, list) is False:
-            raise ValueError(
-                "MergeRegImage requires a list of image names for each image to merge"
-            )
-
-        if isinstance(transform_data, list) is False:
-            raise ValueError(
-                "MergeRegImage requires a list of image transforms for each image to merge"
-            )
-
-        if len(transform_data) != len(self.reg_image.images):
-            raise ValueError(
-                "MergeRegImage number of transforms does not match number of images"
-            )
-
-        def get_transform_data(transform_data):
-            if transform_data is not None:
-                (
-                    itk_composite,
-                    itk_transforms,
-                    final_transform,
-                ) = prepare_wsireg_transform_data(transform_data)
-            else:
-                itk_composite, itk_transforms, final_transform = (
-                    None,
-                    None,
-                    None,
-                )
-
-            return itk_composite, itk_transforms, final_transform
-
-        def prepare_channel_names(sub_image_name, channel_names):
-            return [f"{sub_image_name} - {c}" for c in channel_names]
-
-        self.reg_image.channel_names = [
-            prepare_channel_names(im_name, cnames)
-            for im_name, cnames in zip(
-                sub_image_names, self.reg_image.channel_names
-            )
-        ]
-        self.reg_image.channel_names = [
-            item
-            for sublist in self.reg_image.channel_names
-            for item in sublist
-        ]
-
-        merge_transform_data = [get_transform_data(t) for t in transform_data]
-        composite_transforms = [
-            merge_transform_data[t][0]
-            for t in range(len(merge_transform_data))
-        ]
-        final_transforms = [
-            merge_transform_data[t][2]
-            for t in range(len(merge_transform_data))
-        ]
-
-        spacing_check = [
-            t.output_spacing for t in final_transforms if t is not None
-        ]
-        size_check = [t.output_size for t in final_transforms if t is not None]
-        no_transform_check = [
-            idx for idx, t in enumerate(final_transforms) if t is None
-        ]
-
-        for no_tform in no_transform_check:
-            size_check.extend(
-                [
-                    np.asarray(self.reg_image.images[no_tform].im_dims)[
-                        1:
-                    ].tolist()
-                ]
-            )
-            spacing_check.extend(
-                [
-                    [
-                        self.reg_image.images[no_tform].image_res,
-                        self.reg_image.images[no_tform].image_res,
-                    ]
-                ]
-            )
-
-        if all(spacing_check) is False:
-            raise ValueError(
-                "MergeRegImage all transforms output spacings and untransformed image spacings must match"
-            )
-
-        if all(size_check) is False:
-            raise ValueError(
-                "MergeRegImage all transforms output sizes and untransformed image sizes must match"
-            )
-
-        output_file_name = str(Path(output_dir) / image_name)
-
-        self.prepare_image_info(
-            image_name,
-            final_transform=final_transforms,
-            write_pyramid=write_pyramid,
-            tile_size=tile_size,
-            compression=compression,
-        )
-
-        print(f"saving to {output_file_name}.ome.tiff")
-        with TiffWriter(f"{output_file_name}.ome.tiff", bigtiff=True) as tif:
-            for m_idx, merge_image in enumerate(self.reg_image.images):
-                if self.reg_image.images[m_idx].reader == "sitk":
-                    full_image = sitk.ReadImage(
-                        self.reg_image.images[m_idx].image_filepath
-                    )
-                merge_n_ch = merge_image.n_ch
-                for channel_idx in range(merge_n_ch):
-                    if self.reg_image.images[m_idx].reader != "sitk":
-                        image = self.reg_image.images[
-                            m_idx
-                        ].read_single_channel(channel_idx)
-                        image = np.squeeze(image)
-                        image = sitk.GetImageFromArray(image)
-                        image.SetSpacing(
-                            (
-                                self.reg_image.images[m_idx].image_res,
-                                self.reg_image.images[m_idx].image_res,
-                            )
-                        )
-                    else:
-                        if len(full_image.GetSize()) > 2:
-                            image = full_image[:, :, channel_idx]
-                        else:
-                            image = full_image
-
-                    if composite_transforms[m_idx] is not None:
-                        image = transform_plane(
-                            image,
-                            final_transforms[m_idx],
-                            composite_transforms[m_idx],
-                        )
-
-                    print("saving")
-                    if isinstance(image, sitk.Image):
-                        image = sitk.GetArrayFromImage(image)
-
-                    options = dict(
-                        tile=(tile_size, tile_size),
-                        compression=self.compression,
-                        photometric="rgb"
-                        if merge_image.is_rgb
-                        else "minisblack",
-                        metadata=None,
-                    )
-                    # write OME-XML to the ImageDescription tag of the first page
-                    description = (
-                        self.omexml
-                        if channel_idx == 0 and m_idx == 0
-                        else None
-                    )
-                    # write channel data
-                    print(
-                        f" writing image {m_idx} - channel {channel_idx} - shape: {image.shape}"
-                    )
-                    tif.write(
-                        image,
-                        subifds=self.subifds,
-                        description=description,
-                        **options,
-                    )
-
-                    if write_pyramid:
-                        for pyr_idx in range(1, self.n_pyr_levels):
-                            resize_shape = (
-                                self.pyr_levels[pyr_idx][0],
-                                self.pyr_levels[pyr_idx][1],
-                            )
-                            image = cv2.resize(
-                                image,
-                                resize_shape,
-                                cv2.INTER_LINEAR,
-                            )
-
-                            tif.write(image, **options, subfiletype=1)
-
-            return f"{output_file_name}.ome.tiff"
+        return output_file_name
